refactor: Improve organization of view
