--- lib/src/digit_span_tasks/task_list_page.dart
@@ -1,42 +1,0 @@
-import 'package:flutter/material.dart';
-
-import 'run_dsb.dart';
-import 'run_dsf.dart';
-
-class TaskListPage extends StatefulWidget {
-  const TaskListPage({super.key});
-
-  @override
-  State<TaskListPage> createState() => _TaskListPageState();
-}
-
-class _TaskListPageState extends State<TaskListPage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            ElevatedButton(
-              onPressed: runDigitSpanForward,
-              child: Text(
-                'Digit Span Forward',
-                style: Theme.of(context).textTheme.titleLarge,
-              ),
-            ),
-            const SizedBox(height: 10),
-            ElevatedButton(
-              onPressed: runDigitSpanBackwards,
-              child: Text(
-                'Digit Span Backwards',
-                style: Theme.of(context).textTheme.titleLarge,
-              ),
-            ),
-          ],
-        ),
-      ),
-      backgroundColor: Colors.grey[300],
-    );
-  }
-}
